--- app/controllers/home_controller.rb
@@ -1,4 +1,0 @@
-class HomeController < ApplicationController
-  def index
-  end
-end
