--- app/views/home/index.html.erb
@@ -1,38 +1,0 @@
-<h1>Weather Widget</h1>
-
-<noscript>
-  <h2>Sorry for the inconvenience, but for now, this widget requires JavaScript to be enabled.</h2>
-</noscript>
-
-<div
-  data-action="google-api-callback@window->places#initializeAutocomplete"
-  data-controller="places"
->
-  <%= label_tag "address", "Address" %>
-  <%= search_field_tag "address",
-    nil,
-    data: {
-      action: "change->places#disableSubmit keyup->places#disableSubmit",
-      places_target: "search",
-    },
-    size: 75
-  %>
-  <%= submit_tag "Lookup",
-    data: {
-      action: "click->places#submit",
-      places_target: "submit",
-    },
-    disabled: true
-  %>
-</div>
-
-<details>
-  <summary>Notes</summary>
-  <ul>
-    <li>JavaScript is required to use this widget.</li>
-    <li>This widget is currently limited to locations within the United States.</li>
-    <li>You may enter either a city name or a postal code.</li>
-    <li>Locations must have an associated postal code.</li>
-    <li>Weather data is provided by <%= link_to "OpenWeather", "https://openweathermap.org/", rel: "noopener", target: "_blank" %>.</li>
-  </ul>
-</details>
