--- app/views/locations/edit.html.erb
@@ -1,10 +1,0 @@
-<h1>Editing location</h1>
-
-<%= render "form", location: @location %>
-
-<br>
-
-<div>
-  <%= link_to "Show this location", @location %> |
-  <%= link_to "Back to locations", locations_path %>
-</div>
